move common files into common directory
