+++ src/components/__tests__/Shelf.test.js
@@ -6,7 +6,7 @@ import { shallow, mount } from 'enzyme';
 import Shelf from '../Shelf'
 import BookList from '../BookList'
 import Book from '../Book'
-import { testBooks } from '../../testData'
+import { testBooks } from '../../common/testData'
 
 describe('Shelf', () => {
   let clearQuery, shelf, shelvedBooks, wrapper;
